Fix the scope of pthread variables accross the program
As the threads are created and handled in the main translation unit, there
is no need to make corresponding pthread variables like thread or
condition ids outside the scope of main.c. This also applies to mutex objects
used in the program. One to handle sections of the code in the cpu_monitor.c
module, which must be processed inside the critical section. Another used to
manage threads in main.c depending on the state of the condition variables.

Signed-off-by: Mark Sungurov <mark.sungurov@gmail.com>